refactor: - directory 구조 변경 - zustand 삭제 -> ContextAPI사용 - (auth) directory 생성   - auth 전용 layout 사용
