+++ .env.template
@@ -1,0 +1,3 @@
+NEXT_PUBLIC_JSON_SERVER_URL=http://localhost:8000
+NEXTAUTH_URL=http://localhost:3000
+NEXTAUTH_SECRET=
