+++ package.json
@@ -35,8 +35,7 @@
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
     "uuid": "^9.0.1",
-    "zod": "^3.23.4",
-    "zustand": "^4.5.2"
+    "zod": "^3.23.4"
   },
   "devDependencies": {
     "@types/node": "^20",
