--- src/app/(provider)/(layout)/layout.tsx
@@ -1,16 +1,0 @@
-import MainHeader from "@/components/header";
-
-interface Props {
-  children: React.ReactNode;
-}
-
-function RootLayout({ children }: Props) {
-  return (
-    <>
-      <MainHeader />
-      {children}
-    </>
-  );
-}
-
-export default RootLayout;
