--- src/app/(provider)/(layout)/login/_components/loginPage/LoginPage.tsx
@@ -1,13 +1,0 @@
-import { AuthHeaderTitle } from "@/components/header/title";
-import LoginForm from "@/app/(provider)/(layout)/login/_components/loginForm";
-
-function LoginPage() {
-  return (
-    <div>
-      <AuthHeaderTitle titleName={"로그인"} />
-      <LoginForm />
-    </div>
-  );
-}
-
-export default LoginPage;
