--- src/app/(provider)/(layout)/login/page.tsx
@@ -1,12 +1,0 @@
-import { AuthPage } from "@/components/page";
-import LoginPage from "@/app/(provider)/(layout)/login/_components/loginPage";
-
-function Login() {
-  return (
-    <AuthPage>
-      <LoginPage />
-    </AuthPage>
-  );
-}
-
-export default Login;
