--- src/app/(provider)/(layout)/signup/page.tsx
@@ -1,12 +1,0 @@
-import { AuthPage } from "@/components/page";
-import SignupPage from "@/app/(provider)/(layout)/signup/_components/signupPage";
-
-function SignUp() {
-  return (
-    <AuthPage>
-      <SignupPage />
-    </AuthPage>
-  );
-}
-
-export default SignUp;
