+++ src/app/(providers)/(root)/(auth)/login/page.tsx
@@ -1,0 +1,7 @@
+import LoginPage from "./_components/loginPage";
+
+function Login() {
+  return <LoginPage />;
+}
+
+export default Login;
