+++ src/app/(providers)/(root)/(auth)/signup/_components/formField/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FormField";
