+++ src/app/(providers)/(root)/(auth)/signup/_components/signupForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SignupForm";
