+++ src/app/(providers)/(root)/(auth)/signup/_components/signupPage/index.ts
