+++ src/app/(providers)/(root)/layout.tsx
@@ -1,0 +1,18 @@
+import MainHeader from "@/components/header";
+import { Toaster } from "@/components/ui/toaster";
+
+interface Props {
+  children: React.ReactNode;
+}
+
+function RootLayout({ children }: Props) {
+  return (
+    <>
+      <MainHeader />
+      {children}
+      <Toaster />
+    </>
+  );
+}
+
+export default RootLayout;
