--- src/components/page/AuthPage.tsx
@@ -1,13 +1,0 @@
-interface Props {
-  children: React.ReactNode;
-}
-
-function AuthPage({ children }: Props) {
-  return (
-    <div className="m-auto max-w-[1000px] w-auto min-w-96 flex flex-col gap-5">
-      {children}
-    </div>
-  );
-}
-
-export default AuthPage;
