+++ src/components/page/index.ts
@@ -1,2 +1,1 @@
-export { default as AuthPage } from "./AuthPage";
 export { default as Page } from "./Page";
