--- src/hooks/headerTitle/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useHeaderTitle";
