--- src/provider/rootProvider/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./RootProvider";
