--- src/zustand/config/ccreateSelector.ts
@@ -1,20 +1,0 @@
-import { StoreApi, UseBoundStore } from "zustand";
-
-type WithSelectors<S> = S extends { getState: () => infer T }
-    ? S & { use: { [K in keyof T]: () => T[K] } }
-    : never;
-
-const createSelectors = <S extends UseBoundStore<StoreApi<object>>>(
-    _store: S
-) => {
-  const store = _store as WithSelectors<typeof _store>;
-  store.use = {};
-  Object.keys(store.getState()).forEach((k) => {
-    // eslint-disable-next-line
-    (store.use as any)[k] = () => store((s) => s[k as keyof typeof s]);
-  });
-
-  return store;
-};
-
-export default createSelectors;
