--- src/zustand/config/index.ts
@@ -1,1 +1,0 @@
-export {default} from './ccreateSelector';
