--- src/zustand/stores/signupStepStore/index.ts
@@ -1,1 +1,0 @@
-export {default} from './signupStep.store';
